+++ frontend/src/stores/stateToastList.ts
@@ -1,5 +1,5 @@
 import { atom } from "recoil";
-import { ToastMessage } from "./toastType";
+import { ToastMessage } from "../components/Toast/toastType";
 
 export const toastListState = atom<ToastMessage[]>({
   key: "toastListState",
